making the proper file to run the stuffs
